Imagegen history fixed and tts fixed
Imagegen (gemini-2.5-image) history filtering and tts no audio problems fixed. The issue was duplicate GeminiService file in different file path src/services/geminiservice and services/geminiservice now we use services/geminiservice and redirect traffic from src/services/geminiservice to services/geminiservice. services/geminiservice is now basically just copy paste of the previous src/services/geminiservice. Tested, works.
