+++ metadata.json
@@ -1,5 +1,5 @@
 {
-  "name": "MaestroTutor-GeminiEdition_11_checkpoint",
+  "name": "MaestroTutor-GeminiEdition_12_checkpoint",
   "description": "A fully client-side React application for language learning powered by Google Gemini API. Features real-time voice conversation, image-based context, and personalized tutoring without external backend dependencies.",
   "requestFramePermissions": [
     "camera",
